tools: Faster upgrades to latest versions
- Re-use the tools built from previous docker image
- Move all tools installations into docker volumes for easy migration

+++ scripts/core-utils/build-py-vdev.sh
@@ -10,23 +10,28 @@ set -e
 # in the requirements.txt file in the Gunyah hypervisor source repository
 # in the path hyp/tools/requirements.txt
 
-pushd ${TOOLS_DIR}
+if [[ ! -f ${WORKSPACE}/gunyah-venv/bin/activate ]]; then
 
-echo "Preparing python venv for gunyah"
-python3 -m venv gunyah-venv &&
-    chmod a+x gunyah-venv/bin/activate &&
-    . gunyah-venv/bin/activate &&
-    pip3 install --upgrade pip &&
-    pip3 install wheel &&
-    pip3 install pexpect &&
-    pip3 install lark_parser==0.8.9 &&
-    pip3 install Cheetah3==3.2.6 &&
-    pip3 install pyelftools==0.26 || {
-    echo "Faild to setup python venv"
-    # popd
-    exit 1
-}
+	pushd ${WORKSPACE}
+	echo "Preparing python venv for gunyah"
+	python3 -m venv gunyah-venv &&
+	    chmod a+x gunyah-venv/bin/activate &&
+	    . gunyah-venv/bin/activate &&
+	    pip3 install --upgrade pip &&
+	    pip3 install wheel &&
+	    pip3 install pexpect &&
+	    pip3 install lark_parser==0.8.9 &&
+	    pip3 install Cheetah3==3.2.6 &&
+	    pip3 install pyelftools==0.26 || {
+	    echo "Faild to setup python venv"
+	    # popd
+	    return
+	}
+	popd
 
-popd
-echo "source ${TOOLS_DIR}/gunyah-venv/bin/activate" \
-	>>${TOOLS_DIR}/.tools-env
+	echo ""  >> $WORKSPACE/.wsp-env
+	echo 'source ${WORKSPACE}/gunyah-venv/bin/activate' \
+		 >> $WORKSPACE/.wsp-env
+else
+	echo "Gunyah python venv is already installed"
+fi
