+++ scripts/utils/build-rootfs-img.sh
@@ -27,20 +27,34 @@ set -e
 #                     instead of the whole package
 # ----------------------------------------------------------------------------
 
+IN_ERROR="NO"
 
-if [[ ! -d "${LINUX_DIR}/build" ]]; then
-	echo "Linux build folder not found in ${LINUX_DIR}"
+if [[ -z "${LINUX_DIR}" ]] || [[ ! -f "${LINUX_DIR}/build/arch/arm64/boot/Image" ]]; then
+	echo "Linux build folder not set or kernel image not found in LINUX_DIR : ${LINUX_DIR}"
+	IN_ERROR="YES"
 fi
 
-if [[ -z ${CROSVM_FILE_PATH} ]]; then
-	echo "crosvm file path variable CROSVM_FILE_PATH not found..!! "
-else
-	if [[ ! -f ${CROSVM_FILE_PATH} ]]; then
-		echo "crosvm file (${CROSVM_FILE_PATH}) not found..!! "
-	fi
+if [[ -z "${RAMDISK_FILE_PATH}" ]] || [[ ! -f "${RAMDISK_FILE_PATH}" ]]; then
+	echo "Ramdisk file path is not set or not found RAMDISK_FILE_PATH : ${RAMDISK_FILE_PATH}"
+	IN_ERROR="YES"
 fi
 
-ROOTFS_BASE="${PWD}/rootfs"
+if [[ -z ${CROSVM_FILE_PATH} ]] || [[ ! -f ${CROSVM_FILE_PATH} ]]; then
+	echo "crosvm file path is not set or not found in CROSVM_FILE_PATH : ${CROSVM_FILE_PATH}"
+	IN_ERROR="YES"
+fi
+
+if [[ -z ${WORKSPACE} ]]; then
+	echo "workspace path is not set WORKSPACE : ${WORKSPACE}"
+	IN_ERROR="YES"
+fi
+
+if [[ "$IN_ERROR" = "YES" ]]; then
+	echo "Resolve above errors and run the script again"
+	return
+fi
+
+ROOTFS_BASE="${WORKSPACE}/rootfs"
 ROOTFS_REFERENCE_DIR="${ROOTFS_BASE}/reference"
 
 mkdir -p ${ROOTFS_REFERENCE_DIR}
@@ -59,39 +73,49 @@ cd ${ROOTFS_BASE}
 echo "Now preparing Linaro stock rootfs image"
 
 ROOTFS_LINARO_STOCK="${ROOTFS_BASE}/linaro-stock"
-rm -rf ${ROOTFS_LINARO_STOCK}
+LINARO_ROOTFS_IMAGE_FILE_NAME=initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4
+LINARO_ROOTFS_IMAGE=${LINARO_ROOTFS_IMAGE_FILE_NAME}.gz
+LINARO_ROOTGS_URL=https://snapshots.linaro.org/member-builds/qcomlt/testimages/arm64/1379
 
-mkdir -p ${ROOTFS_LINARO_STOCK}
+if [[ -d ${ROOTFS_REFERENCE_DIR}/etc/systemd/system ]]; then
+	echo "Reference folder already exists in ${ROOTFS_REFERENCE_DIR}"
+else
+	mkdir -p ${ROOTFS_LINARO_STOCK}
 
-cd ${ROOTFS_LINARO_STOCK}
+	cd ${ROOTFS_LINARO_STOCK}
 
-echo "Now downloading Linaro reference rootfs image"
+	echo "Now downloading Linaro reference rootfs image"
 
-# Download initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4 from
-# linaro website
-wget https://snapshots.linaro.org/member-builds/qcomlt/testimages/arm64/1379/initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4.gz
+	# Download the rootfs image $LINARO_ROOTFS_IMAGE from linaro website
+	if [[ ! -f  ${ROOTFS_LINARO_STOCK}/${LINARO_ROOTFS_IMAGE_FILE_NAME} ]]; then
+		wget ${LINARO_ROOTGS_URL}/${LINARO_ROOTFS_IMAGE}
 
-echo "Download completed, decompressing the image"
+		echo "Download completed, decompressing the image"
 
-# Decompress the image
-gunzip initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4.gz
+		# Decompress the image LINARO_ROOTFS_IMAGE as LINARO_ROOTFS_IMAGE_FILE_NAME
+		gunzip ${LINARO_ROOTFS_IMAGE}
+	fi
 
-# It would be nice if resize works, but newer e2fsck is needed TBD later!!
-#resize2fs initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4 512M
-#e2fsck -f initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4
+	# It would be nice if resize works, but newer e2fsck is needed TBD later!!
+	#resize2fs initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4 512M
+	#e2fsck -f initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4
 
-echo "Decompression completed, mount the image to ${ROOTFS_LINARO_STOCK}/mnt"
+	echo "Decompression completed, mount the image to ${ROOTFS_LINARO_STOCK}/mnt"
 
-mkdir -p ${ROOTFS_LINARO_STOCK}/mnt
+	mkdir -p ${ROOTFS_LINARO_STOCK}/mnt
 
-# mount the Linaro stock rootfs image to extract all the files
-sudo mount -o loop initramfs-tiny-image-qemuarm64-20230321073831-1379.rootfs.ext4  ${ROOTFS_LINARO_STOCK}/mnt
+	# mount the Linaro stock rootfs image to extract all the files
+	sudo mount -o loop ${LINARO_ROOTFS_IMAGE_FILE_NAME}  ${ROOTFS_LINARO_STOCK}/mnt
 
-echo "Copy the file tree to reference tree"
+	echo "Copy the file tree to reference tree"
 
-sudo cp -r -v -p ${ROOTFS_LINARO_STOCK}/mnt/*   ${ROOTFS_REFERENCE_DIR}
+	sudo cp -r -v -p ${ROOTFS_LINARO_STOCK}/mnt/*   ${ROOTFS_REFERENCE_DIR}
 
-sudo umount ${ROOTFS_LINARO_STOCK}/mnt
+	sudo umount ${ROOTFS_LINARO_STOCK}/mnt
+
+	# Retain if needed later
+	#rm -rf ${ROOTFS_LINARO_STOCK}
+fi
 
 # ----------------------------------------------------------------------------
 # Linux kernel built modules
@@ -101,19 +125,22 @@ sudo umount ${ROOTFS_LINARO_STOCK}/mnt
 UTS_RELEASE=`cat ${LINUX_DIR}/build/include/config/kernel.release`
 KO_FILES_DST="${ROOTFS_REFERENCE_DIR}/lib/modules/${UTS_RELEASE}/kernel"
 
-echo "Copying the linux ko files to destination dir ${KO_FILES_DST}"
-cd ${LINUX_DIR}/build
+if [[ -d ${KO_FILES_DST} ]]; then
+	echo "Kernel files already copied to reference destination"
+else
+	echo "Copying the linux ko files to destination dir ${KO_FILES_DST}"
+	cd ${LINUX_DIR}/build
 
-for f in $(find . -iname "*.ko");
-do
-	DST_DIR=$(dirname "${KO_FILES_DST}/$f")
-	#echo "mkdir ${DST_DIR}"
-	sudo mkdir -p ${DST_DIR}
-	sudo cp -v -p $f ${DST_DIR}
-done
+	for f in $(find . -iname "*.ko");
+	do
+		DST_DIR=$(dirname "${KO_FILES_DST}/$f")
+		#echo "mkdir ${DST_DIR}"
+		sudo mkdir -p ${DST_DIR}
+		sudo cp -v -p $f ${DST_DIR}
+	done
 
-echo "Done copying linux kernel object files to reference rootfs tree"
-
+	echo "Done copying linux kernel object files to reference rootfs tree"
+fi
 cd ${ROOTFS_BASE}
 
 # ----------------------------------------------------------------------------
@@ -123,27 +150,37 @@ cd ${ROOTFS_BASE}
 # these include, crosvm binary, SVM linux kernel image, ramdisk
 
 SVM_DESTINATION=${ROOTFS_REFERENCE_DIR}/usr/gunyah
-sudo mkdir -p ${SVM_DESTINATION}
 
-if [[ ! -z ${CROSVM_FILE_PATH} ]]; then
-	if [[ -f ${CROSVM_FILE_PATH} ]]; then
+if [[ -f ${SVM_DESTINATION}/$(basename ${CROSVM_FILE_PATH}) ]]; then
+	echo "Crosvm file is already copied to reference folder"
+else
+	sudo mkdir -p ${SVM_DESTINATION}
+	if [[ ! -z ${CROSVM_FILE_PATH} ]] && [[ -f ${CROSVM_FILE_PATH} ]]; then
 		echo "Copying crosvm file to rootfs reference tree"
 		sudo cp -v -p ${CROSVM_FILE_PATH} ${SVM_DESTINATION}
 	fi
 fi
 
-sudo cp -v -p ${LINUX_DIR}/build/arch/arm64/boot/Image ${SVM_DESTINATION}
-sudo cp -v -p ${RAMDISK_FILE_PATH} ${SVM_DESTINATION}
+if [[ ! -f ${SVM_DESTINATION}/Image ]]; then
+	sudo cp -v -p ${LINUX_DIR}/build/arch/arm64/boot/Image ${SVM_DESTINATION}
+fi
 
-echo -e '#!/bin/sh\n\n/usr/gunyah/crosvm --no-syslog run --disable-sandbox \\'  > ./svm.sh
-echo -e '--serial=type=stdout,hardware=virtio-console,console,stdin,num=1 \\' >> ./svm.sh
-echo -e '--serial=type=stdout,hardware=serial,earlycon,num=1 \\' >> ./svm.sh
-echo -e '--initrd /usr/gunyah/initrd.img  --no-balloon --no-rng \\' >> ./svm.sh
-echo -e '--params "rw root=/dev/ram rdinit=/sbin/init earlyprintk=serial panic=0" \\' >> ./svm.sh
-echo -e ' /usr/gunyah/Image $@\n' >> ./svm.sh
+if [[ ! -f ${SVM_DESTINATION}/$(basename ${RAMDISK_FILE_PATH}) ]]; then
+	sudo cp -v -p ${RAMDISK_FILE_PATH} ${SVM_DESTINATION}
+fi
 
-sudo cp ./svm.sh ${SVM_DESTINATION}
-sudo chmod 0775 ${SVM_DESTINATION}/svm.sh
+if [[ ! -f ${SVM_DESTINATION}/svm.sh ]]; then
+	echo -e '#!/bin/sh\n\n/usr/gunyah/crosvm --no-syslog run --disable-sandbox \\'  > ./svm.sh
+	echo -e '--serial=type=stdout,hardware=virtio-console,console,stdin,num=1 \\' >> ./svm.sh
+	echo -e '--serial=type=stdout,hardware=serial,earlycon,num=1 \\' >> ./svm.sh
+	echo -e '--initrd /usr/gunyah/initrd.img  --no-balloon --no-rng \\' >> ./svm.sh
+	echo -e '--params "rw root=/dev/ram rdinit=/sbin/init earlyprintk=serial panic=0" \\' >> ./svm.sh
+	echo -e ' /usr/gunyah/Image $@\n' >> ./svm.sh
+
+	sudo cp ./svm.sh ${SVM_DESTINATION}
+	rm -f ./svm.sh
+	sudo chmod 0775 ${SVM_DESTINATION}/svm.sh
+fi
 
 echo "Completed copying crosvm and SVM kernel files to rootfs reference tree"
 
@@ -158,65 +195,79 @@ echo "Completed copying crosvm and SVM kernel files to rootfs reference tree"
 # Following Reference commands are derived from files fetch.log, build.log at
 # https://snapshots.linaro.org/member-builds/qcomlt/testimages/arm64/1379/
 
-if [[ ! -f ~/bin/repo ]]; then
-	echo "Installing repo into local bin folder"
-        mkdir -p ~/bin
-        #export PATH=~/bin:$PATH
-	#echo "$PATH"
-        curl http://commondatastorage.googleapis.com/git-repo-downloads/repo > ~/bin/repo
-        chmod a+x ~/bin/repo
+if [[ ! -f ${ROOTFS_REFERENCE_DIR}/lib/libgcc_s.so.1 ]]; then
+	if [[ ! -f ~/bin/repo ]]; then
+		echo "Installing repo into local bin folder"
+		mkdir -p ~/bin
+		#export PATH=~/bin:$PATH
+		#echo "$PATH"
+		curl http://commondatastorage.googleapis.com/git-repo-downloads/repo > ~/bin/repo
+		chmod a+x ~/bin/repo
+	fi
+
+	if [[ ! -f ~/.gitconfig ]]; then
+		echo "Warning: setting the git global config to default values..!!"
+		git config --global user.name "$USER"
+		git config --global user.email "$USER@local.com"
+		git config --global color.ui auto
+	fi
+
+	ROOTFS_IMAGE_TO_BUILD="rpb-console-image"
+	export MACHINE=qemuarm64
+	export DISTRO=rpb
+
+	mkdir ${ROOTFS_BASE}/oe-rpb
+	cd ${ROOTFS_BASE}/oe-rpb
+
+	# fetch
+	~/bin/repo init -u https://github.com/96boards/oe-rpb-manifest.git -b qcom/master
+	~/bin/repo sync
+
+	# add config for libgcc and other virtualization options
+	echo -e "\n" > ./extra_local.conf
+	echo "INHERIT += 'buildstats buildstats-summary'" >> ./extra_local.conf
+	echo "PREFERRED_PROVIDER_virtual/kernel = 'linux-dummy'" >> ./extra_local.conf
+	echo "PREFERRED_PROVIDER_android-tools-conf = 'android-tools-conf-configfs'" >> ./extra_local.conf
+	echo "CORE_IMAGE_EXTRA_INSTALL += 'openssh libgcc'" >> ./extra_local.conf
+	echo "PACKAGE_INSTALL:append = ' openssh libgcc'" >> ./extra_local.conf
+	echo "DISTRO_FEATURES:append = ' virtualization'" >> ./extra_local.conf
+	echo -e "\n" >> ./extra_local.conf
+
+	echo -e "\n\n" > ./bblayers.conf
+
+
+	source setup-environment build
+
+	cat ../extra_local.conf >> conf/local.conf
+	cat ../bblayers.conf >> conf/bblayers.conf
+
+	echo '"Dumping local.conf.."'
+	cat conf/local.conf
+
+	bitbake -e > bitbake-environment
+
+	bitbake ${ROOTFS_IMAGE_TO_BUILD}
+
+	# Completed the build. The file libgcc_s.so.1 should be available at path
+	# ${ROOTFS_BASE}/oe-rpb/build/tmp-rpb-glibc/sysroots-components/cortexa57/libgcc/usr/lib/libgcc_s.so.1
+	LIBGCC_OUT_PATH="build/tmp-rpb-glibc/sysroots-components/cortexa57/libgcc/usr/lib"
+
+	sudo cp ${ROOTFS_BASE}/oe-rpb/${LIBGCC_OUT_PATH}/libgcc_s.so.1 ${ROOTFS_REFERENCE_DIR}/lib
+	sudo chmod 0755 ${ROOTFS_REFERENCE_DIR}/lib/libgcc_s.so.1
+
+	rm -rf ${ROOTFS_BASE}/oe-rpb
 fi
 
-if [[ ! -f ~/.gitconfig ]]; then
-	echo "Warning: setting the git global config to default values..!!"
-	git config --global user.name "$USER"
-	git config --global user.email "$USER@local.com"
-	git config --global color.ui auto
-fi
-
-ROOTFS_IMAGE_TO_BUILD="rpb-console-image"
-export MACHINE=qemuarm64
-export DISTRO=rpb
-
-mkdir oe-rpb
-cd oe-rpb
-
-# fetch
-~/bin/repo init -u https://github.com/96boards/oe-rpb-manifest.git -b qcom/master
-~/bin/repo sync
-
-# add config for libgcc and other virtualization options
-echo -e "\n" > ./extra_local.conf
-echo "INHERIT += 'buildstats buildstats-summary'" >> ./extra_local.conf
-echo "PREFERRED_PROVIDER_virtual/kernel = 'linux-dummy'" >> ./extra_local.conf
-echo "PREFERRED_PROVIDER_android-tools-conf = 'android-tools-conf-configfs'" >> ./extra_local.conf
-echo "CORE_IMAGE_EXTRA_INSTALL += 'openssh libgcc'" >> ./extra_local.conf
-echo "PACKAGE_INSTALL:append = ' openssh libgcc'" >> ./extra_local.conf
-echo "DISTRO_FEATURES:append = ' virtualization'" >> ./extra_local.conf
-echo -e "\n" >> ./extra_local.conf
-
-echo -e "\n\n" > ./bblayers.conf
-
-
-source setup-environment build
-
-cat ../extra_local.conf >> conf/local.conf
-cat ../bblayers.conf >> conf/bblayers.conf
-
-echo '"Dumping local.conf.."'
-cat conf/local.conf
-
-bitbake -e > bitbake-environment
-
-bitbake ${ROOTFS_IMAGE_TO_BUILD}
-
-# Completed the build. The file libgcc_s.so.1 should be available at path
-# ${ROOTFS_BASE}/oe-rpb/build/tmp-rpb-glibc/sysroots-components/cortexa57/libgcc/usr/lib/libgcc_s.so.1
-LIBGCC_OUT_PATH="build/tmp-rpb-glibc/sysroots-components/cortexa57/libgcc/usr/lib"
-sudo cp ${ROOTFS_BASE}/oe-rpb/${LIBGCC_OUT_PATH}/libgcc_s.so.1 ${ROOTFS_REFERENCE_DIR}/lib
-sudo chmod 0755 ${ROOTFS_REFERENCE_DIR}/lib/libgcc_s.so.1
+echo "Successfully created the reference files folder for rootfs image at : `pwd`"
 
 # -----------------------------------------------------------------------------
 # Create a extfs device image of required size
 
-# ~/utils/bldextfs.sh -f ~/rootfs/rootfs/reference -o rootfs.extfs.img -s 1G
+if [[ -f ${WORKSPACE}/rootfs/rootfs-extfs-disk.img ]]; then
+	echo "Rootfs image already exists, delete this file if need to create"
+	echo "  new file with any modified content from reference folder"
+else
+	echo "Creating rootfs image file from reference : `pwd`"
+	cd ${WORKSPACE}/rootfs
+	. ~/utils/bldextfs.sh -f ${WORKSPACE}/rootfs/reference -o ${WORKSPACE}/rootfs/rootfs-extfs-disk.img -s 800M
+fi
